feat(uptime): Add function to determine which partitions a checker owns
This adds `determine_owned_partitions`, which produces a set of partition numbers that this checker owns based on total checkers, total partitions and this checker's checker number.

Simplifies the interface to calling run_config_provider since we can determine partitions via config rather than passing in.

+++ src/app/config.rs
@@ -84,7 +84,7 @@ pub struct Config {
 
     /// How many config partitions do we want to keep in redis? We shouldn't change this once
     /// assigned unless we plan a migration process.
-    pub config_provider_redis_total_partitions: u64,
+    pub config_provider_redis_total_partitions: u16,
 
     /// The general purpose redis node to use with this service
     pub redis_host: String,
@@ -99,7 +99,7 @@ pub struct Config {
     pub checker_id: String,
 
     /// Total number of uptime checkers running
-    pub total_checkers: usize,
+    pub total_checkers: u16,
 }
 
 impl Default for Config {
